+++ src/cli/exportGraph.test.js
@@ -6,36 +6,44 @@ import {run} from "./testUtil";
 import {help, makeExportGraph} from "./exportGraph";
 import {Graph, NodeAddress, EdgeAddress} from "../core/graph";
 import type {IAnalysisAdapter} from "../analysis/analysisAdapter";
+import stringify from "json-stable-stringify";
 
 import * as RepoIdRegistry from "../core/repoIdRegistry";
 import {makeRepoId, type RepoId} from "../core/repoId";
 
-const makeMockDeclaration = () => ({
-  name: "mock",
-  nodePrefix: NodeAddress.empty,
-  edgePrefix: EdgeAddress.empty,
-  nodeTypes: [],
-  edgeTypes: [],
-});
-
 class MockAnalysisAdapter implements IAnalysisAdapter {
-  _resolutionGraph: Graph | null;
+  _resolutionGraph: ?Graph;
+  _name: string;
 
-  declaration() {
-    return makeMockDeclaration();
+  /**
+   * Takes a name for the plugin, and a graph that
+   * is provided as a result of a successful load.
+   * If no graph is provided, then load will fail.
+   */
+  constructor(name: string, resolutionGraph: ?Graph) {
+    this._name = name;
+    this._resolutionGraph = resolutionGraph;
   }
 
-  load(
+  declaration() {
+    return {
+      name: this._name,
+      nodePrefix: NodeAddress.empty,
+      edgePrefix: EdgeAddress.empty,
+      nodeTypes: [],
+      edgeTypes: [],
+    };
+  }
+
+  async load(
     _unused_sourcecredDirectory: string,
     _unused_repoId: RepoId
   ): Promise<Graph> {
-    return new Promise((resolve, reject) => {
-      if (this._resolutionGraph != null) {
-        resolve(this._resolutionGraph);
-      } else {
-        reject("MockAnalysisAdapterRejects");
-      }
-    });
+    if (this._resolutionGraph != null) {
+      return this._resolutionGraph;
+    } else {
+      throw new Error("MockAnalysisAdapterRejects");
+    }
   }
 }
 
@@ -62,7 +70,7 @@ describe("cli/exportGraph", () => {
   });
 
   describe("'exportGraph' command", () => {
-    function setupRegistryWithId(repoId: RepoId) {
+    function setUpRegistryWithId(repoId: RepoId) {
       const dirname = tmp.dirSync().name;
       process.env.SOURCECRED_DIRECTORY = dirname;
       const registry = RepoIdRegistry.addEntry(RepoIdRegistry.emptyRegistry(), {
@@ -73,13 +81,25 @@ describe("cli/exportGraph", () => {
     }
 
     it("prints usage with '--help'", async () => {
-      const exportGraph = makeExportGraph([new MockAnalysisAdapter()]);
+      const exportGraph = makeExportGraph([new MockAnalysisAdapter("foo")]);
       expect(await run(exportGraph, ["--help"])).toEqual({
         exitCode: 0,
         stdout: expect.arrayContaining([
           expect.stringMatching(/^usage: sourcecred export-graph/),
         ]),
         stderr: [],
+      });
+    });
+
+    it("errors if no repoId is provided", async () => {
+      const exportGraph = makeExportGraph([new MockAnalysisAdapter("foo")]);
+      expect(await run(exportGraph, [])).toEqual({
+        exitCode: 1,
+        stdout: [],
+        stderr: expect.arrayContaining([
+          "fatal: no repository ID provided",
+          "fatal: run 'sourcecred help export-graph' for help",
+        ]),
       });
     });
 
@@ -96,15 +116,14 @@ describe("cli/exportGraph", () => {
     });
 
     it("prints json-serialized graph to stdout for a single plugin", async () => {
-      const mockAdapter = new MockAnalysisAdapter();
-      const exportGraph = makeExportGraph([mockAdapter]);
       const g = new Graph().addNode(NodeAddress.empty);
-      mockAdapter._resolutionGraph = g;
-      setupRegistryWithId(makeRepoId("foo", "bar"));
+      const mockAdapter = new MockAnalysisAdapter("foo", g);
+      const exportGraph = makeExportGraph([mockAdapter]);
+      setUpRegistryWithId(makeRepoId("foo", "bar"));
       const result = run(exportGraph, ["foo/bar"]);
       expect(await result).toEqual({
         exitCode: 0,
-        stdout: JSON.stringify(g.toJSON()).split("\n"),
+        stdout: [stringify(g.toJSON())],
         stderr: [],
       });
     });
@@ -112,73 +131,70 @@ describe("cli/exportGraph", () => {
     it("merges graphs for multiple plugins", async () => {
       const g1 = new Graph().addNode(NodeAddress.fromParts(["g1"]));
       const g2 = new Graph().addNode(NodeAddress.fromParts(["g2"]));
-      const m1 = new MockAnalysisAdapter();
-      m1._resolutionGraph = g1;
-      const m2 = new MockAnalysisAdapter();
-      m2._resolutionGraph = g2;
+      const m1 = new MockAnalysisAdapter("foo", g1);
+      const m2 = new MockAnalysisAdapter("bar", g2);
       const mergedGraph = Graph.merge([g1, g2]);
-      setupRegistryWithId(makeRepoId("foo", "bar"));
+      setUpRegistryWithId(makeRepoId("foo", "bar"));
       const exportGraph = makeExportGraph([m1, m2]);
       expect(await run(exportGraph, ["foo/bar"])).toEqual({
         exitCode: 0,
-        stdout: JSON.stringify(mergedGraph.toJSON()).split("\n"),
+        stdout: [stringify(mergedGraph.toJSON())],
         stderr: [],
       });
     });
 
     it("errors if multiple repos are provided", async () => {
-      const m1 = new MockAnalysisAdapter();
-      const m2 = new MockAnalysisAdapter();
+      const m1 = new MockAnalysisAdapter("foo");
+      const m2 = new MockAnalysisAdapter("bar");
       const exportGraph = makeExportGraph([m1, m2]);
       expect(await run(exportGraph, ["foo/bar", "zod/zoink"])).toEqual({
         exitCode: 1,
         stdout: [],
         stderr: [
-          "fatal: multiple repoIds provided",
+          "fatal: multiple repository IDs provided",
           "fatal: run 'sourcecred help export-graph' for help",
         ],
       });
     });
 
     it("errors if the repoId was not loaded first", async () => {
-      const mockAdapter = new MockAnalysisAdapter();
-      const exportGraph = makeExportGraph([mockAdapter]);
       const g = new Graph().addNode(NodeAddress.empty);
-      mockAdapter._resolutionGraph = g;
-      setupRegistryWithId(makeRepoId("foo", "bar"));
+      const mockAdapter = new MockAnalysisAdapter("mock", g);
+      const exportGraph = makeExportGraph([mockAdapter]);
+      setUpRegistryWithId(makeRepoId("foo", "bar"));
       const result = run(exportGraph, ["zod/zoink"]);
       expect(await result).toEqual({
         exitCode: 1,
         stdout: [],
         stderr: [
-          "fatal: repoId zod/zoink not loaded",
-          "try running `sourcecred load zod/zoink` first.",
+          "fatal: repository ID zod/zoink not loaded",
+          "Try running `sourcecred load zod/zoink` first.",
         ],
       });
     });
 
     it("passes the right arguments to adapter.load", async () => {
-      const mockAdapter = new MockAnalysisAdapter();
+      const mockAdapter = new MockAnalysisAdapter("zoo");
       const exportGraph = makeExportGraph([mockAdapter]);
       const repoId = makeRepoId("foo", "bar");
       // $ExpectFlowError
       mockAdapter.load = jest.fn();
-      const directory = setupRegistryWithId(repoId);
+      const directory = setUpRegistryWithId(repoId);
       await run(exportGraph, ["foo/bar"]);
       expect(mockAdapter.load).toHaveBeenCalledWith(directory, repoId);
     });
 
     it("reports the failing plugin when a plugin rejects", async () => {
-      const mockAdapter = new MockAnalysisAdapter();
+      const mockAdapter = new MockAnalysisAdapter("bar");
       const exportGraph = makeExportGraph([mockAdapter]);
       const repoId = makeRepoId("foo", "bar");
-      setupRegistryWithId(repoId);
+      setUpRegistryWithId(repoId);
       const result = await run(exportGraph, ["foo/bar"]);
       expect(result).toEqual({
         exitCode: 1,
         stdout: [],
         stderr: [
-          'fatal: plugin "mock" errored: MockAnalysisAdapterRejects',
+          'fatal: plugin "bar" errored: MockAnalysisAdapterRejects',
           "fatal: run 'sourcecred help export-graph' for help",
         ],
       });